change in User account Api

+++ QuizAPI/Quiz.Web.APIModel/Quiz/QuizApiModel.cs
@@ -11,8 +11,8 @@ namespace Quiz.Web.APIModel.Quiz
         public int Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
-        public int Duration { get; set; }
-        public int PassingPercentage { get; set; }
+        public decimal Duration { get; set; }
+        public decimal PassingPercentage { get; set; }
 
         public bool IsActive { get; set; } = true;
